chore: transformed docs application into standalone bootstrap app and removed redundant module
