+++ apps/docs/src/app/app.component.ts
@@ -1,9 +1,12 @@
 import { Component, OnInit } from '@angular/core';
+import { RouterOutlet } from '@angular/router';
 import { ThemingService } from '@fundamental-ngx/core/theming';
 
 @Component({
     selector: 'app-root',
-    template: `<router-outlet></router-outlet>`
+    template: ` <router-outlet></router-outlet>`,
+    imports: [RouterOutlet],
+    standalone: true
 })
 export class AppComponent implements OnInit {
     /** @hidden */
