+++ libs/cdk/src/lib/utils/directives/clicked/index.ts
@@ -1,3 +1,4 @@
 export * from './clicked-behavior.module';
 export * from './clicked.directive';
 export * from './fdk-clicked-provider.service';
+export * from './provide-fdk-clicked';
