+++ libs/core/src/lib/theming/index.ts
@@ -1,7 +1,8 @@
-export * from './theming.module';
-export * from './tokens';
 export * from './config';
 export * from './interfaces/theme.interface';
 export * from './interfaces/theming-config.interface';
+export * from './provide-theming';
 export * from './standard-themes';
+export * from './theming.module';
 export * from './theming.service';
+export * from './tokens';
