+++ libs/core/src/lib/theming/tokens.ts
@@ -1,4 +1,7 @@
 import { InjectionToken } from '@angular/core';
+import { BaseThemingConfig } from './config';
 import { ThemingConfig } from './interfaces/theming-config.interface';
 
-export const THEMING_CONFIG_TOKEN = new InjectionToken<Partial<ThemingConfig>>('FdThemingConfig');
+export const THEMING_CONFIG_TOKEN = new InjectionToken<Partial<ThemingConfig>>('FdThemingConfig', {
+    factory: () => new BaseThemingConfig()
+});
