+++ libs/docs/shared/src/lib/utilities/libraries.ts
@@ -1,5 +1,5 @@
 import { InjectionToken } from '@angular/core';
 
-export type Libraries = 'core' | 'platform' | 'datetime-adapter' | 'cx';
+export type Libraries = 'cdk' | 'core' | 'platform' | 'datetime-adapter' | 'cx';
 
 export const CURRENT_LIB = new InjectionToken<Libraries>('Current Library');
